CPR-329 Re-organise processors directory
