+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/message/processors/delius/OffenderCreatedEventProcessor.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.personrecord.message.listeners.processors
+package uk.gov.justice.digital.hmpps.personrecord.message.processors.delius
 
 import kotlinx.coroutines.runBlocking
 import org.slf4j.LoggerFactory
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Component
 import uk.gov.justice.digital.hmpps.personrecord.client.OffenderDetailRestClient
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.DeliusOffenderDetail
 import uk.gov.justice.digital.hmpps.personrecord.jpa.repository.PersonRepository
+import uk.gov.justice.digital.hmpps.personrecord.message.processors.EventProcessor
 import uk.gov.justice.digital.hmpps.personrecord.model.DomainEvent
 import uk.gov.justice.digital.hmpps.personrecord.model.Person
 import uk.gov.justice.digital.hmpps.personrecord.service.PersonService
